4.21, 4.22

+++ week4/bloglist/src/routes/blog.ts
@@ -1,10 +1,10 @@
-import express from "express"
+import express, {Request} from "express"
 import { Blog } from "../models/blog"
 import { User } from "../models/user"
 import { BlogType } from "../types"
 import { validateBlog } from "../utils/blog_validator"
 import jwt from "jsonwebtoken"
-import { CustomRequest } from "../utils/tokenProcessing"
+import { CustomRequest, userExtractor } from "../utils/tokenProcessing"
 
 const blogRouter = express.Router()
 
@@ -13,23 +13,17 @@ blogRouter.get("/", async (_request, response) => {
   return response.json(blogs)
 })
 
-blogRouter.post("/", async (request, response) => {
+blogRouter.post("/", userExtractor, async (request, response) => {
   const blogValidate = validateBlog(request.body)
   if (!blogValidate) return response.status(400).end()
 
-  const decodedToken = jwt.verify(
-    (request as CustomRequest).token,
-    process.env.SECRET as string
-  )
 
-  if (!decodedToken || typeof decodedToken === "string" || !decodedToken.id)
-    return response.status(401).json({ error: "token missing or invalid" })
 
-  const user = await User.findById(decodedToken.id)
+  const user = await User.findById((request as CustomRequest).user)
 
   if (!user) return response.status(400).json({ error: "User not found." })
 
-  const blog = new Blog(blogValidate)
+  const blog = new Blog({ ...blogValidate, user: (request as CustomRequest).user })
 
   const savedBlog = await blog.save()
 
@@ -39,9 +33,16 @@ blogRouter.post("/", async (request, response) => {
   return response.status(201).json(savedBlog)
 })
 
-blogRouter.delete("/:id", async (request, response) => {
+blogRouter.delete("/:id", userExtractor, async (request: Request, response) => {
   const idToBeDeleted: string | undefined = request.params.id
   if (!idToBeDeleted) return response.status(400).end()
+
+  const blog = await Blog.findById(idToBeDeleted)
+  
+  if (!blog) return response.status(400).end()
+
+  if (blog.user.toString() !== (request as CustomRequest).user)
+    return response.status(401).json({ error: "Unauthorized" })  
 
   const result: BlogType | undefined | null = await Blog.findByIdAndDelete(
     idToBeDeleted
